- Prettify error messages - Fix info button css - Raise error on compact flag
