+++ redisinsight/ui/src/packages/redisgraph/src/styles/styles.less
@@ -75,6 +75,13 @@
             justify-content: space-between;
             padding-top: 12px;
             padding-left: 12px;
+
+            .euiButtonIcon {
+                position: relative;
+	              top: 0px !important;
+	              right:6px !important;
+            }
+
         }
 
         .info-props {
@@ -226,3 +233,9 @@
     margin-right: 12px !important;
 }
 
+
+.responseFail {
+    color: red;
+    padding: 12px !important;
+    font-family: monospace !important;
+}
